refactor(direct-control): extract restart begin result type
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves Civ7RestartAndBeginResult into src/setup/restart.ts while keeping public facade type re-exports in index.ts.

Preserves restart/begin result shape and does not change lifecycle orchestration, command routing, App UI snapshot parsing, Tuner readiness waiting, restart-output validation, or no-replay behavior. Remaining facade-owned health result types, public procedure schemas, operation/proof telemetry, hotseat runtime proof, AI ingestion, CLI semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance remain pending.

Validation:
- bun run --cwd packages/civ7-direct-control test -- restart-lifecycle.test.ts setup-and-lifecycle.test.ts public-api.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- git diff --cached --check
- ownership scan for restart/begin result type definition

Local package/type relocation proof only; no runtime, live-game, AI-on-hotseat, telemetry, semantic CLI, or procedure-core proof claimed.
